Move file stuff to a separate context file
